+++ perl/FSM/Support/ISFPublicInterfaceContract.pm
@@ -433,6 +433,7 @@ sub build_isf_public_interface_contract {
             't/1264-isf-enum-member-set-expression-values.t',
             't/1265-isf-enum-member-switch-branch-values.t',
             't/1266-isf-enum-member-drive-values.t',
+            't/1267-isf-enum-member-drive-call-values.t',
         ],
         guidance => [
             'Treat this as the first bounded public ISF downstream-consumer contract, advertised through embedding.isf_public_interface.',